requirements.txt
fastapi
uvicorn
python-telegram-bot==20.7
sqlalchemy
psycopg2-binary
python-dotenv
mysql-connector-python